# openapi/intel.yaml
openapi: 3.0.3
info:
  title: VMAN - INTEL Service API
  version: 0.1.0
  description: |
    REST API for the INTEL service of VMAN. Manages VM templates, VM lifecycle, and disks.
servers:
  - url: http://localhost:8000
    description: Local development server
...
paths:
  /templates:
    post:
      tags: [templates]
      summary: Create a VM template
      operationId: createTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VMTemplateCreate'
      responses:
        '201':
          description: Template created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VMTemplate'
        '400':
          description: Invalid input
    get:
      tags: [templates]
      summary: List templates
      operationId: listTemplates
      parameters: []
      responses:
        '200':
          description: List of templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VMTemplate'
  /templates/{name}:
    delete:
      tags: [templates]
      summary: Delete a VM template by name
      operationId: deleteTemplate
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Template deleted
        '404':
          description: Template not found
...
components:
  schemas:
    VMTemplate:
      type: object
      properties:
        name:
          type: string
        cpu_count:
          type: integer
          minimum: 1
        ram_amount:
          type: integer
          minimum: 1
      required: [name, cpu_count, ram_amount]
    VMTemplateCreate:
      allOf:
        - $ref: '#/components/schemas/VMTemplate'
...
security: []